+++ src/Symfony/Component/Security/Core/Role/RoleHierarchy.php
@@ -36,6 +36,8 @@ class RoleHierarchy implements RoleHierarchyInterface
      */
     public function getReachableRoles(array $roles)
     {
+        @trigger_error(sprintf('The %s() method is deprecated since Symfony 4.3 and will be removed in 5.0. Use roles as strings and the getReachableRoleNames() method instead.', __METHOD__), E_USER_DEPRECATED);
+
         $reachableRoles = $roles;
         foreach ($roles as $role) {
             if (!isset($this->map[$role->getRole()])) {
@@ -44,6 +46,37 @@ class RoleHierarchy implements RoleHierarchyInterface
 
             foreach ($this->map[$role->getRole()] as $r) {
                 $reachableRoles[] = new Role($r);
+            }
+        }
+
+        return $reachableRoles;
+    }
+
+    /**
+     * @param string[] $roles
+     *
+     * @return string[]
+     */
+    public function getReachableRoleNames(array $roles): array
+    {
+        $reachableRoles = $roles = array_map(
+            function ($role) {
+                if ($role instanceof Role) {
+                    return $role->getRole();
+                }
+
+                return $role;
+            },
+            $roles
+        );
+
+        foreach ($roles as $role) {
+            if (!isset($this->map[$role])) {
+                continue;
+            }
+
+            foreach ($this->map[$role] as $r) {
+                $reachableRoles[] = $r;
             }
         }