+++ src/Symfony/Component/Security/Core/Tests/Authorization/Voter/RoleVoterTest.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Security\Core\Tests\Authorization\Voter;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Security\Core\Authentication\Token\AbstractToken;
 use Symfony\Component\Security\Core\Authorization\Voter\RoleVoter;
 use Symfony\Component\Security\Core\Authorization\Voter\VoterInterface;
 use Symfony\Component\Security\Core\Role\Role;
@@ -19,6 +20,7 @@ use Symfony\Component\Security\Core\Role\Role;
 class RoleVoterTest extends TestCase
 {
     /**
+     * @group legacy
      * @dataProvider getVoteTests
      */
     public function testVote($roles, $attributes, $expected)
@@ -26,6 +28,16 @@ class RoleVoterTest extends TestCase
         $voter = new RoleVoter();
 
         $this->assertSame($expected, $voter->vote($this->getToken($roles), null, $attributes));
+    }
+
+    /**
+     * @dataProvider getVoteTests
+     */
+    public function testVoteUsingTokenThatReturnsRoleNames($roles, $attributes, $expected)
+    {
+        $voter = new RoleVoter();
+
+        $this->assertSame($expected, $voter->vote($this->getTokenWithRoleNames($roles), null, $attributes));
     }
 
     public function getVoteTests()
@@ -41,6 +53,23 @@ class RoleVoterTest extends TestCase
             // Test mixed Types
             [[], [[]], VoterInterface::ACCESS_ABSTAIN],
             [[], [new \stdClass()], VoterInterface::ACCESS_ABSTAIN],
+        ];
+    }
+
+    /**
+     * @group legacy
+     * @dataProvider getLegacyVoteOnRoleObjectsTests
+     */
+    public function testVoteOnRoleObjects($roles, $attributes, $expected)
+    {
+        $voter = new RoleVoter();
+
+        $this->assertSame($expected, $voter->vote($this->getToken($roles), null, $attributes));
+    }
+
+    public function getLegacyVoteOnRoleObjectsTests()
+    {
+        return [
             [['ROLE_BAR'], [new Role('ROLE_BAR')], VoterInterface::ACCESS_GRANTED],
             [['ROLE_BAR'], [new Role('ROLE_FOO')], VoterInterface::ACCESS_DENIED],
         ];
@@ -54,6 +83,16 @@ class RoleVoterTest extends TestCase
         $token = $this->getMockBuilder('Symfony\Component\Security\Core\Authentication\Token\TokenInterface')->getMock();
         $token->expects($this->once())
               ->method('getRoles')
+              ->will($this->returnValue($roles));
+
+        return $token;
+    }
+
+    protected function getTokenWithRoleNames(array $roles)
+    {
+        $token = $this->getMockBuilder(AbstractToken::class)->getMock();
+        $token->expects($this->once())
+              ->method('getRoleNames')
               ->will($this->returnValue($roles));
 
         return $token;